Move SmokeInvocationTraceContext into SmokeOperationsHTTP1 library.
